Refactored to starta and stop telegram application based on env profile
